requirements.txt
streamlit>=1.20
tensorflow>=2.9
Pillow>=9.0
numpy>=1.21